--- 06_control_and_form/src/130_reminder/start/Todo.jsx
@@ -1,16 +1,0 @@
-const Todo = () => {
-  const todosList = [
-    {
-      id: 1,
-      content: "店予約する",
-    },
-    {
-      id: 2,
-      content: "卵買う",
-    },
-    {
-      id: 3,
-      content: "郵便出す",
-    },
-  ];
-};
