+++ 06_control_and_form/src/130_reminder/start/comonents/Todo.jsx
@@ -1,0 +1,42 @@
+import List from "./List";
+import Form from "./Form";
+
+import { useState } from "react";
+
+const Todo = () => {
+  const todosList = [
+    {
+      id: 1,
+      content: "店予約する",
+    },
+    {
+      id: 2,
+      content: "卵買う",
+    },
+    {
+      id: 3,
+      content: "郵便出す",
+    },
+  ];
+
+  const [todos, setTodos] = useState(todosList);
+
+  const deleteTodo = (id) => {
+    const newTodos = todos.filter((todo) => {
+      return todo.id !== id;
+    });
+    setTodos(newTodos);
+  };
+
+  const createTodo = (todo) => {
+    setTodos([...todos, todo]);
+  };
+  return (
+    <>
+      <List todos={todos} deleteTodo={deleteTodo} />
+      <Form createTodo={createTodo} />
+    </>
+  );
+};
+
+export default Todo;
